# Conflicts:
#   addons/custom/forlife_purchase/views/purchase_order_view.xml

+++ addons/custom/nhanh_connector/models/product_product.py
@@ -9,15 +9,13 @@ import json
 _logger = logging.getLogger(__name__)
 
 
-class ProductNhanh(models.Model):
-    _inherit = 'product.template'
+class ProductProduct(models.Model):
+    _inherit = 'product.product'
 
-    nhanh_id = fields.Integer(string="Id Nhanh.Vn")
-    code_product = fields.Char(string="Mã sản phẩm")
-    ## Nếu tạo sản phẩm từ Odoo == True else == False
+    nhanh_id = fields.Integer(string="Id Nhanh.Vn", copy=False)
     check_data_odoo = fields.Boolean(string='Check dữ liệu từ odoo or Nhanh', default=True)
-    width_product = fields.Float('Width')
-    height_product = fields.Float('Height')
+    width_product = fields.Float('Width', copy=False)
+    height_product = fields.Float('Height', copy=False)
 
     @api.model
     def create(self, vals):
@@ -31,12 +29,13 @@ class ProductNhanh(models.Model):
         if res.check_data_odoo and res.brand_id.id:
             nhanh_configs = constant.get_nhanh_configs(self, brand_ids=[res.brand_id.id]).get(res.brand_id.id)
             if nhanh_configs.get('nhanh_connector.nhanh_app_id', '') or nhanh_configs.get(
-                    'nhanh_connector.nhanh_business_id', '') or nhanh_configs.get('nhanh_connector.nhanh_access_token', ''):
+                    'nhanh_connector.nhanh_business_id', '') or nhanh_configs.get('nhanh_connector.nhanh_access_token',
+                                                                                  ''):
 
                 data = [{
                     "id": res.id,
                     "name": res.name,
-                    "code": res.code_product,
+                    "code": res.barcode if res.barcode else '',
                     "barcode": res.barcode if res.barcode else '',
                     "importPrice": res.list_price,
                     "price": res.list_price,
@@ -77,36 +76,51 @@ class ProductNhanh(models.Model):
 
     def write(self, vals):
         res = super().write(vals)
-        if 'name' not in vals and 'code_product' not in vals and 'barcode' not in vals and 'list_price' not in vals and 'weight' not in vals:
+        if 'name' not in vals and 'barcode' not in vals and 'list_price' not in vals and 'weight' not in vals:
             return res
+        data = []
         for item in self:
-            data = [{
-                "id": item.id,
+            if not item.nhanh_id:
+                continue
+            data.append({
+                "id": item._origin.id,
                 "idNhanh": item.nhanh_id,
                 "name": item.name,
-                "code": item.code_product,
+                "code": item.barcode if item.barcode else '',
                 "barcode": item.barcode if item.barcode else '',
                 "importPrice": item.list_price,
                 "price": item.list_price,
                 "shippingWeight": item.weight * 1000,
                 "status": 'New'
-            }]
-            self.synchronized_price_nhanh(data)
+            })
+        if not data:
+            return res
+        self.synchronized_price_nhanh(data)
         return res
 
     def unlink(self):
+        data = []
         for item in self:
-            data = '[{"id": "' + str(item.id) + '","idNhanh":"' + str(item.nhanh_id) + '", "price": "' + str(int(
-                item.list_price)) + '", "name": "' + str(item.name) + '", "shippingWeight": "' + str(
-                int(item.weight)) + '", "status": "' + 'Inactive' + '", "barcode": "' + str(
-                item.barcode if item.barcode else '') + '"}]'
-            self.synchronized_price_nhanh(data)
-        res = super().unlink()
-
-        return res
+            if not item.nhanh_id:
+                continue
+            data.append({
+                "id": item._origin.id,
+                "idNhanh": item.nhanh_id,
+                "name": item.name,
+                "code": item.barcode if item.barcode else '',
+                "barcode": item.barcode if item.barcode else '',
+                "importPrice": item.list_price,
+                "price": item.list_price,
+                "shippingWeight": item.weight * 1000,
+                "status": 'Inactive'
+            })
+        if not data:
+            return super().unlink()
+        self.synchronized_price_nhanh(data)
+        return super().unlink()
 
     def synchronized_price_nhanh(self, data):
-        nhanh_configs = constant.get_nhanh_configs(self, brand_ids=[self.brand_id.id])
+        nhanh_configs = constant.get_nhanh_configs(self, brand_ids=[self.brand_id.id]).get(self.brand_id.id)
         if nhanh_configs.get('nhanh_connector.nhanh_app_id', '') or nhanh_configs.get(
                 'nhanh_connector.nhanh_business_id', '') or nhanh_configs.get('nhanh_connector.nhanh_access_token', ''):
             status_nhanh = 1
